+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ### Unreleased
 ---
+misc
+---
+- Remove rumqttd-old in favour of rumqttd (#530)
+
 rumqttc
 -------
 - Add support for native-tls within rumqttc (#501)
@@ -8,6 +12,7 @@ rumqttc
 
 rumqttd
 -------
+- Make router::Meter public (#521)
 - Add meters related to router, subscriptions, and connections (#508)
 - Allow multi-tenancy validation for mtls clients with `Org` set in certificates (#505)
 - Add `tracing` for structured, context-aware logging (#499, #503)